requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.1
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2==2.9.5
python-decouple==3.6
pytz==2022.6
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.2
